fixed RegNext(Vec). Not having init was giving problems earlier... also did some directory refactoring
