+++ .gitignore
@@ -6,6 +6,7 @@
 javacore*
 target/
 .idea/
+project/
